Add Room, Repo, ViewModel in HomeFragment

+++ app/src/main/java/com/example/todoapp/Model/Subtask.kt
@@ -1,9 +1,10 @@
 package com.example.todoapp.Model
 
+import java.io.Serializable
 import java.util.Date
 
 data class Subtask(
-    var id : Int,
+    var id : Long,
     var title: String,
     var isFinish : Boolean = false
 )
